% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{addData}
\alias{addData}
\title{\code{\link{addData}} : add a data set, e.g., calculated ratios}
\usage{
addData(data, ID, dat, col, processing, replace = FALSE)
}
\arguments{
\item{data}{the current platexpress data set}

\item{ID}{the ID of the new data}

\item{dat}{the new data, must be a matrix akin to other data in the set}

\item{col}{plot color for the new data, an RGB string w/o alpha suffix}

\item{processing}{optional processing note}

\item{replace}{replace existing data, default: FALSE}
}
\description{
\code{\link{addData}} : add a data set, e.g., calculated ratios
}
\seealso{
\code{\link{readPlateData}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
