% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{addModel.multiple_fits}
\alias{addModel.multiple_fits}
\title{Add results from  package \pkg{growthrates}
to \code{platexpress} object.}
\usage{
\method{addModel}{multiple_fits}(fit, data, ID = "model", ...)
}
\arguments{
\item{fit}{a \code{\link[grofit:gcFit]{gcFit}} object}

\item{data}{\code{platexpress} data object}

\item{ID}{data ID for the new object}

\item{...}{arguments passed to \code{\link{addData}}}
}
\description{
Calls the appropriate \code{\link[stats:predict]{predict}} methods
for the growth curves fits returned by
by \pkg{growthrates}'s batch functions (`all_`),
and adds results to the \code{platexpress} data object.
}
