% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{addModel_dpseg}
\alias{addModel_dpseg}
\title{Add results from  \code{\link[dpseg:dpseg]{dpseg}}
to \code{platexpress} object.}
\usage{
addModel_dpseg(data, segments, ID = "y", add.slopes = FALSE, ...)
}
\arguments{
\item{data}{\code{platexpress} data object}

\item{segments}{\code{\link[dpseg:dpseg]{dpseg}} object}

\item{ID}{data ID for the new object}

\item{add.slopes}{add slopes instead of reconstructed data}

\item{...}{arguments passed to \code{\link{addModel}}}
}
\description{
Calls the \code{\link[dpseg:predict.dpseg]{predict.dpseg}} method
to reconstruct the piecewise linear growth curve, and adds it to the
\code{platexpress} data object. Option \code{add.slopes}
allows to add a time-course of growth rates (slopes
of the linear segments) instead.
}
