% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{addModel_segmented}
\alias{addModel_segmented}
\title{Add results from  \code{\link{segmented_plate}}
to \code{platexpress} object.}
\usage{
addModel_segmented(data, fit, ID = "y", add.slopes = FALSE, ...)
}
\arguments{
\item{data}{\code{platexpress} data object}

\item{fit}{list of \code{\link[segmented:segmented]{segmented}} objects}

\item{ID}{data ID for the new object}

\item{add.slopes}{add slopes instead of reconstructed data}

\item{...}{arguments passed to \code{\link{addModel}}}
}
\description{
Calls the \code{\link[segmented:predict.segmented]{predict.segmented}} method
to reconstruct the piecewise linear growth curve, and adds it to the
\code{platexpress} data object. Option \code{add.slopes}
allows to add a time-course of growth rates (slopes
of the linear segments) instead.
}
