% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{amountColors}
\alias{amountColors}
\title{Add Amount Colors}
\usage{
amountColors(
  map,
  substance = "substance",
  amount = "amount",
  color = "color",
  colf = colorRamps::matlab.like2
)
}
\arguments{
\item{map}{the plate map, see  \code{\link{readPlateMap}}}

\item{substance}{name of the substance column in \code{map}}

\item{amount}{name of the amount column in \code{map}}

\item{color}{name of the color column to be added to \code{map}}

\item{colf}{color palette function, default is colorRamp's
\code{\link[colorRamps:matlab.like2]{matlab.like2}}}
}
\description{
add a color palette to the plate layout map, with colors along the range
of added amounts of a given substance. Substance and amount columns are eg.
auto-generated by \code{\link{readPlateMap}} with options \code{asep}
and \code{afields}.
}
\seealso{
\code{\link{readPlateMap}}, \code{\link{replaceAmounts}}
}
