% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\docType{data}
\name{ap12data}
\alias{ap12data}
\title{ap12data: example data by Dennis Dienst and Alice Pawloski, incl. the
plate reader measurements of E.coli growth, expressing a fluorescent
proteins, in a Synergy Mx platereader}
\format{a list of time-courses of absorbance and fluorescence data, read in by readPlateData("AP12.csv", type="Synergy", data.ids=c("600","YFP_50:500,535")) and readPlateMap("AP12_layout.csv", fields=c("strain","samples"))}
\usage{
data(ap12data)
}
\description{
\itemize{
  \item Data:
  \item Time: the interpolated 'master' time
  \item Temperature: the temperature time-course
  \item Data matrix '600': well absorbance at 600 nm, i.e., the OD,
  \item Data matrix 'YFP_50:500,535': the YFP fluorescence measured by excitation at 500 nm and emission at 535 nm 
  \item Plate Layout:
  \item The plate layout table indicates the different strains tested, biological replicates (B1 to B3), and blank wells (containing only growth medium) 
}
}
\seealso{
\code{\link{readPlateData}} and \code{\link{readPlateMap}}
}
\keyword{datasets}

