% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\docType{data}
\name{ap12plate}
\alias{ap12plate}
\title{ap12plate. example plate layout map by Dennis Dienst and Alice Pawloski,
fitting to the experimental data in \code{\link{ap12data}}
The plate layout table indicates the different strains tested, biological
replicates (B1 to B3), and blank wells (containing only growth medium)}
\description{
ap12plate. example plate layout map by Dennis Dienst and Alice Pawloski,
fitting to the experimental data in \code{\link{ap12data}}
The plate layout table indicates the different strains tested, biological
replicates (B1 to B3), and blank wells (containing only growth medium)
}
