% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{bandwidthCV.2}
\alias{bandwidthCV.2}
\title{hack of cellGrowth bandwidthCV to work with platexpress data format
TODO: select bandwidths from data$Time}
\usage{
bandwidthCV.2(data, did, mid, wells, bandwidths = seq(0.5, 10, length.out =
  30), nFold = 10, nWell = 50, cutoff = 0.95, scaleY = identity)
}
\description{
hack of cellGrowth bandwidthCV to work with platexpress data format
TODO: select bandwidths from data$Time
}

