% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{boxData}
\alias{boxData}
\title{returns data for group of wells in a given range of the x-axis}
\usage{
boxData(data, rng, groups, mid, did = "OD", stat = FALSE, plot = TRUE,
  type = "box", etype = "ci")
}
\arguments{
\item{data}{\code{\link{platexpress}} data, see \code{\link{readPlateData}}}

\item{rng}{x-axis range or a single point, for the latter the closest
point will be selected}

\item{groups}{a grouping of wells, see \code{\link{getGroups}}}

\item{mid}{the x-axis ID, if multiple x-axes are present}

\item{did}{the y-axis data to be grouped}

\item{stat}{if FALSE (default) the full grouped data will be returned as
a list (which can be used for \code{boxplot} or statistics afterwards),
if TRUE the means of the data will be returned (as plotted)}

\item{plot}{if TRUE a box-plot or bar-plot will be plotted}

\item{type}{either "box" (default) or "bar" for box-plot or bar-plot}

\item{etype}{type of statistics to be used for error bars in the bar-plot,
either "ci" (default) for the 95%-confidence interval or "se" for
the standard error}
}
\description{
returns data for group of wells in a given range of the x-axis
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}

