% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R, R/interfaces.R.bak.R
\name{callGrofit}
\alias{callGrofit}
\alias{callGrofit}
\title{high-level wrapper for \code{\link[grofit:grofit]{grofit}}}
\usage{
callGrofit(data, plate, yid = "OD", fields = c("strain", "medium",
  "substance"), control, model.type = c("richards", "gompertz.exp",
  "gompertz"), nboot.gc = 100, plot = TRUE, interactive = FALSE, ...)

callGrofit(data, plate, yid = "OD", fields = c("strain", "medium",
  "substance"), control, model.type = c("richards", "gompertz.exp",
  "gompertz"), nboot.gc = 100, plot = TRUE, interactive = FALSE, ...)
}
\arguments{
\item{data}{a platexpress data set, see \code{\link{readPlateData}}}

\item{plate}{plate layout map, see \code{\link{readPlateMap}}, columns
of this map can be converted to \code{\link[grofit:grofit]{grofit}} data
annotation}

\item{yid}{data ID of the data to be converted; default
is to take the first of \code{data$dataIDs}}

\item{fields}{column IDs in the plate layout map to be used for
\code{\link[grofit:grofit]{grofit}} data annotation; if \code{plate} is not
missing at least 1 column is required}

\item{control}{an object of class \code{grofit.control} with an additional
parameter \code{plot}, as provided by the \code{platexpress} function
\code{\link{grofit.2.control}}; if present arguments \code{model.type},
\code{nboot.gc} and \code{plot} will be ignored (unless the latter
is not present)}

\item{model.type}{models that \code{grofit} will attempt to fit,
see \code{\link[grofit:grofit.control]{grofit.control}}; will be igonred
if \code{control} is passed directly}

\item{nboot.gc}{numbers of bootstrap samples,
see \code{\link[grofit:grofit.control]{grofit.control}}; will be igonred
if \code{control} is passed directly}

\item{plot}{set to TRUE for plots even without \code{interactive} use}

\item{interactive}{set to TRUE for interactive growth curve fitting}

\item{...}{further arguments to \code{\link{grofit.2.control}}}

\item{amount}{name of a column in \code{plate} that provides a numeric
'amount', to be used as a 'dose' annotation in grofit}

\item{data}{a platexpress data set, see \code{\link{readPlateData}}}

\item{plate}{plate layout map, see \code{\link{readPlateMap}}, columns
of this map can be converted to \code{\link[grofit:grofit]{grofit}} data
annotation}

\item{yid}{data ID of the data to be converted, from \code{data$dataIDs}}

\item{fields}{column IDs in the plate layout map to be used for
\code{\link[grofit:grofit]{grofit}} data annotation; if \code{plate} is not
missing at least 1 column is required}

\item{control}{an object of class \code{grofit.control} with an additional
parameter \code{plot}, as provided by the \code{platexpress} function
\code{\link{grofit.2.control}}; if present arguments \code{model.type},
\code{nboot.gc} and \code{plot} will be ignored (unless the latter
is not present)}

\item{model.type}{models that \code{grofit} will attempt to fit,
see \code{\link[grofit:grofit.control]{grofit.control}}; will be igonred
if \code{control} is passed directly}

\item{nboot.gc}{numbers of bootstrap samples,
see \code{\link[grofit:grofit.control]{grofit.control}}; will be igonred
if \code{control} is passed directly}

\item{plot}{set to TRUE for plots even without \code{interactive} use}

\item{interactive}{set to TRUE for interactive growth curve fitting}

\item{...}{further arguments to \code{\link{grofit.2.control}}}
}
\description{
Calls \code{\link[grofit:grofit]{grofit}} directly from \code{platexpress}
plate data and layout, using established settings. Please see
the documentation of package \code{\link[grofit:grofit]{grofit}}
for details of the fitting procedure. This high-level wrapper used
\code{platexpress} functions \code{\link{data2grofit}},
\code{\link{grofit.2.control}}, \code{\link{gcFit.2}} to convert
data, set parameters and call \code{\link[grofit:grofit]{grofit}},
and then maps results (lag phase lambda, growth rate mu and capacity A)
to the \code{plate} layout map, and \code{\link{addModel}} to add modelled
prediction of the fitted data to the plate \code{data} object.
The function returns a list containing the new \code{data} and \code{plate}
objects, as well as the original \code{grofit} result object.

Calls \code{\link[grofit:grofit]{grofit}} directly from \code{platexpress}
plate data and layout, using established settings. Please see
the documentation of package \code{\link[grofit:grofit]{grofit}}
for details of the fitting procedure. This high-level wrapper used
\code{platexpress} functions \code{\link{data2grofit}},
\code{\link{grofit.2.control}}, \code{\link{gcFit.2}} to convert
data, set parameters and call \code{\link[grofit:grofit]{grofit}},
and then maps results (lag phase lambda, growth rate mu and capacity A)
to the \code{plate} layout map, and \code{\link{addModel}} to add modelled
prediction of the fitted data to the plate \code{data} object.
The function returns a list containing the new \code{data} and \code{plate}
objects, as well as the original \code{grofit} result object.
}
