% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{data2grofit}
\alias{data2grofit}
\title{\code{\link{data2grofit}} : converts \code{package:platexpress} data to
\code{package:grofit} data format}
\usage{
data2grofit(data, did = "OD", min.time, max.time, wells, plate, eid, dose)
}
\arguments{
\item{data}{the current platexpress data set, see \code{\link{readPlateData}}}

\item{did}{data ID of the data to be converted, from \code{data$dataIDs}}

\item{max.time}{maximal time of the data to be used}

\item{wells}{column IDs of the data set to use, if missing all wells
are taken}

\item{plate}{plate layout map, see \code{\link{readPlateMap}}, columns
of this map can be converted to \code{package:grofit} data annotation}

\item{eid}{column IDs in the plate layout map to be used for
\code{package:groFit} data annotation; if missing but \code{plate} is
present, the columns 2 and 3 are used}

\item{dose}{vector of doses in each well, used as the third column of
\code{package:grofit} data annotation, where it can be used for dose-response
calculations}
}
\description{
\code{\link{data2grofit}} : converts \code{package:platexpress} data to
\code{package:grofit} data format
}
\details{
Returns a simple list with two entries \code{time} and \code{data},
as required for \code{package:grofit}.
}
\examples{
grdat <- data2grofit(data)
fit <- gcFit(grdat$time, grdat$data)
}

