% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{doseResponse}
\alias{doseResponse}
\title{Dose-Response Plots with Error Bars}
\usage{
doseResponse(map, wells, val, amount = "amount", substance = "substance",
  col = "black", pch = 1, bartype = "range", barl = 0.05, all = FALSE,
  line = TRUE, na.y = 0, add = FALSE, ylim, xlim, ylab, xlab,
  verb = TRUE, ...)
}
\arguments{
\item{map}{a plate layout map with columns \code{amount} and some
calculated value in column \code{val},
eg. results from \pkg{grofit} or \pkg{growthrates}}

\item{wells}{a list of well IDs to be used in the plot}

\item{val}{the name of a column in \code{map} containing numeric values
that should be plotted on y-axis}

\item{amount}{the name of a column in \code{map} providing numeric values
that should be plotted on x-axis, typically a substance added to wells
in multiple replicates; the default value "amount" is automatically
generated from appropriate plate layout files by \code{link{readPlateMap}}}

\item{substance}{the name of a column in \code{map} providing the names of
the substance in \code{amount}, used as x-axis label; if it doesn't
correspond to a column, its value is directly used as label}

\item{col}{either a valid color representation or the name of a column in
\code{map} providing colors for each well or a, if \code{wells} is a named
list of wells, a named vector of colors with matching names. NOTE for the
second case, that wells with the same \code{amount} should have the same color, only
the first color for a given \code{amount} is used; such colors are automatically
assigned from a color ramp mapped to the numerical range of \code{amount} by
\code{\link{readPlateMap}}, see \code{\link{amountColors}};}

\item{pch}{pch of the mean value points}

\item{bartype}{type of the error bars, "range" for the full range
of the data, where the average value will be the
\code{\link[stats:median]{median}}; or "ci95" for the 95\% confidence
interval, "sd" for the standard deviation, or "se" for the standard error
where the average value will be the \code{\link[base:mean]{mean}};
TODO: implement boxplot-like quantiles}

\item{barl}{length of the horizontal lines of error bars (argument \code{length}
of function \code{\link{arrows}})}

\item{all}{logical indicating whether to plot all data points additionally to
error bars}

\item{line}{logical indicating whether to plot a line connecting the mean
values}

\item{na.y}{value to be used to plot replicates with \code{NA} in
column \code{val}; set to NA to supress plotting}

\item{add}{add data to existing plot}

\item{ylim}{limits of the y-axis}

\item{xlim}{limits of the x-axis}

\item{ylab}{alternative label for the y-axis, default is to use argument \code{val}}

\item{xlab}{alternative label for the x-axis, default is to use argument
\code{substance}, or if this is a column in \code{map}, the substance
indicated there}

\item{verb}{print progress messages}

\item{...}{arguments passed on to the main setup \code{\link{plot}}}
}
\description{
plots a continuous value for each well as a function
of an \code{amount} of a \code{substance} and generates error bars
for all replicates at a given amount; returns the plotted mean values
and error bar ranges.
}
