% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{dpseg_plate}
\alias{dpseg_plate}
\title{Call \code{\link[dpseg:dpseg]{dpseg}} for selected wells.}
\usage{
dpseg_plate(data, yid = "OD", wells, log = TRUE, xid, verb = 0, ...)
}
\arguments{
\item{data}{\code{platexpress} data object}

\item{yid}{ID of the \code{platexpress} data to use}

\item{wells}{subset and plot order of wells}

\item{log}{use ln of the data}

\item{xid}{x-axis data ID in \code{platexpress} data}

\item{verb}{progress messages}

\item{...}{arguments passed to \code{\link[dpseg:dpseg]{dpseg}}}
}
\description{
The algorithm in \code{\link[dpseg:dpseg]{dpseg}} splits curves
into linear segments.  If the passed data is a biomass measure
(eg. OD), and option \code{log=TRUE} the slopes correspond to local
growth rate (per time unit).
}
