% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{findWavelength}
\alias{findWavelength}
\title{\code{\link{findWavelength}}:
a color selector for \code{\link{showSpectrum}}; expects the user to
click on the spectrum, then draws a vertical line at the clicked
wavelength using \code{\link{plotWavelength}} and records the wavelength
in nm.}
\usage{
findWavelength(n = 1, ...)
}
\arguments{
\item{n}{number of clicks to record}

\item{...}{further arguments to \code{\link{plotWavelength}}, for
selecting plot symbol and text size, and positions}
}
\description{
\code{\link{findWavelength}}:
a color selector for \code{\link{showSpectrum}}; expects the user to
click on the spectrum, then draws a vertical line at the clicked
wavelength using \code{\link{plotWavelength}} and records the wavelength
in nm.
}
\seealso{
\code{\link{showSpectrum}}, \code{\link{plotWavelength}}, \code{\link{wavelength2RGB}}
}

