% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R, R/interfaces.R.bak.R
\name{fitCellGrowths.2}
\alias{fitCellGrowths.2}
\alias{fitCellGrowths.2}
\title{batch wrapper for fitCellGrowth
TODO: align in/out with gcFit.2 or optionally cellGrowth::fitCellGrowths}
\usage{
fitCellGrowths.2(data, yid, mid, wells, ...)

fitCellGrowths.2(data, yid, mid, wells, ...)
}
\arguments{
\item{data}{platexpress data set, see \code{\link{readPlateData}}}

\item{yid}{data ID of the data to be converted for grofit, from
\code{data$dataIDs}}

\item{mid}{the x-axis to be used, defaults to the first available
(usually "Time")}

\item{wells}{column IDs of the data set to use, if missing all wells
are used}

\item{...}{further parameters to the \code{cellGrowth} package function
\code{\link[cellGrowth:fitCellGrowth]{fitCellGrowth}}}

\item{xid}{the x-axis to be used, defaults to the first available
(usually "Time")}

\item{data}{platexpress data set, see \code{\link{readPlateData}}}

\item{yid}{data ID of the data to be converted for grofit, from
\code{data$dataIDs}}

\item{wells}{column IDs of the data set to use, if missing all wells
are used}

\item{...}{further parameters to the \code{cellGrowth} package function
\code{\link[cellGrowth:fitCellGrowth]{fitCellGrowth}}}
}
\description{
batch wrapper for fitCellGrowth
TODO: align in/out with gcFit.2 or optionally cellGrowth::fitCellGrowths

batch wrapper for fitCellGrowth
TODO: align in/out with gcFit.2 or optionally cellGrowth::fitCellGrowths
}
\seealso{
\code{\link[cellGrowth:fitCellGrowth]{fitCellGrowth}}

\code{\link[cellGrowth:fitCellGrowth]{fitCellGrowth}}
}
