% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{getWells}
\alias{getWells}
\title{get a filtered list of wells}
\usage{
getWells(plate, blanks = FALSE, values)
}
\arguments{
\item{plate}{the plate layout map, see \code{\link{readPlateMap}}}

\item{blanks}{if set to \code{TRUE} (default) the argument \code{values}
is optional, and only blank values will be returned.}

\item{values}{a named list of strings, key:value pairs, where the
list names (keys) correspond to column names in the plate layout map
and the values are entries in these columns}
}
\value{
Returns the list of wells according to argument \code{values},
or a list of blanks if \code{blanks==TRUE}
}
\description{
get a filtered list of wells
}
