% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{grofit_plate}
\alias{grofit_plate}
\title{high-level wrapper for \pkg{grofit}}
\usage{
grofit_plate(data, plate, yid, amount, fields = c("strain", "medium",
  "substance"), control, model.type = c("richards", "logistic",
  "gompertz.exp", "gompertz"), nboot.gc = 100, plot = TRUE,
  interactive = FALSE, verb = TRUE, col = "#0000FF", ...)
}
\arguments{
\item{data}{a platexpress data set, see \code{\link{readPlateData}}}

\item{plate}{plate layout map, see \code{\link{readPlateMap}}, columns
of this map can be converted to \pkg{grofit} data
annotation, and results will be returned in the same order of wells}

\item{yid}{data ID of the data to be converted; default
is to take the first of \code{data$dataIDs}}

\item{amount}{name of a column in \code{plate} that provides a numeric
'amount', to be used as a 'dose' annotation in grofit}

\item{fields}{column IDs in the plate layout map to be used for
\pkg{grofit} data annotation; if \code{plate} is not
missing at least 1 column is required}

\item{control}{an object of class \code{grofit.control} with an additional
parameter \code{plot}, as provided by the \code{platexpress} function
\code{\link{grofit.2.control}}; if present arguments \code{model.type},
\code{nboot.gc} and \code{plot} will be ignored (unless the latter
is not present)}

\item{model.type}{models that \code{grofit} will attempt to fit,
see \code{\link[grofit:grofit.control]{grofit.control}}; will be igonred
if \code{control} is passed directly}

\item{nboot.gc}{numbers of bootstrap samples,
see \code{\link[grofit:grofit.control]{grofit.control}}; will be igonred
if \code{control} is passed directly}

\item{plot}{set to TRUE for plots even without \code{interactive} use}

\item{interactive}{set to TRUE for interactive growth curve fitting}

\item{verb}{print messages}

\item{col}{color of the model data in the plate data object}

\item{...}{further arguments to \code{\link{grofit.2.control}}}
}
\description{
Calls \pkg{grofit} directly from \code{platexpress}
plate data and layout, using established settings. Please see
the documentation of package \pkg{grofit}
for details of the fitting procedure. This high-level wrapper uses
\code{platexpress} functions \code{\link{data2grofit}},
\code{\link{grofit.2.control}}, \code{\link{gcFit.2}} to convert
data, to set parameters and to call \pkg{grofit}.
}
\details{
The function returns a list of fits, as returnted by
\code{platexpress}'s \code{\link{gcFit.2}}, a copy of
grofit's \code{\link[grofit:gcFit]{gcFit}} with modified
plotting and interactive behaviour but otherwise identical.
All run parameters of \code{\link[grofit:gcFit]{gcFit}}
can be set via option \code{control}, where \code{platexpress}'
\code{\link{grofit.2.control}} add the differing plot and interactive
settings.
}
