% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R, R/interfaces.R.bak.R
\name{grofitResults}
\alias{grofitResults}
\alias{grofitResults}
\title{parse \code{\link[grofit:grofit]{grofit}} results}
\usage{
grofitResults(fit)

grofitResults(fit)
}
\arguments{
\item{fit}{grofit object, the result of a call to
\code{\link[grofit:gcFit]{gcFit}}
or the \code{platexpress} version \code{\link{gcFit.2}}}

\item{fit}{grofit object, the result of a call to
\code{\link[grofit:gcFit]{gcFit}}
or the \code{platexpress} version \code{\link{gcFit.2}}}
}
\description{
parses the output of \code{\link{gcFit.2}} into a table
of the main model parameters, for each well

parses the output of \code{\link{gcFit.2}} into a table
of the main model parameters, for each well
}
\seealso{
\code{\link{data2grofit}}, \code{\link{growthratesResults}}

\code{\link{data2grofit}}, \code{\link{growthratesResults}}
}
