% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{groupStats}
\alias{groupStats}
\title{Group Statistics}
\usage{
groupStats(data, groups, yids, xid)
}
\arguments{
\item{data}{\code{\link{platexpress}} data, see \code{\link{readPlateData}}}

\item{groups}{a list of well grouped wells, as produced by
\code{\link{getGroups}}}

\item{yids}{data IDs for which statistics should be reported,
if missing stats for all data will be reported}

\item{xid}{ID of a data-set in the input data that can be used as x-axis
instead of the default Time vector}
}
\value{
Returns a data structure similar to the input data, but
where actual data is replaced  by statistics over grouped wells.
}
\description{
calculates simple statistics for groups as plotted in
\code{\link{viewGroups}}. TODO: actually use these stats
in viewGroups, optionally add group stats to original data structure
}
\details{
Calculates the simple statistics over grouped wells
(means, 95% confidence intervals, stdandard errors) along the x-axis
(usually time).
}
\seealso{
\code{\link{readPlateMap}}, \code{\link{viewGroups}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
