% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeModels.R
\name{ode_anacat}
\alias{ode_anacat}
\alias{grow_anacat}
\title{Anabolism vs. Catabolism Growth Model}
\usage{
ode_anacat(time, init, parms, ...)

grow_anacat(time, parms, ...)
}
\arguments{
\item{time}{actual time (for the ode) resp. vector of simulation time steps.}

\item{init}{named vector with the initial state of the system}

\item{parms}{parameters of the anabolism vs. catabolism growth model
\itemize{
   \item \code{y} initial biomass concentration (cell number),
   \item \code{s} initial substrate concentration
   \item \code{mumax} maximum growth rate (1/time),
   \item \code{K} saturation or half-maximal growth constant
   \item \code{Y} the yield factor (fraction of substrate converted to biomass
   \item \code{phi} culture dilution rate
   \item \code{sin} substrate concentration in medium influx
}}

\item{\dots}{placeholder for additional parameters (for user-extended versions of this function)}
}
\value{
For \code{ode_anacata}: matrix containing the simulation outputs.
The return value of has also class \code{deSolve}.

For \code{grow_anacata}: vector of dependent variable (\code{y}) and
  its log-transformed values (\code{log_y}):

\itemize{
\item \code{time} time of the simulation
\item \code{y} cell concentration
}
}
\description{
Anabolism vs. Catabolism Growth Model

Simple anacat growth model
}
\details{
Function \code{ode_anacata} is the system of differential equations,
whereas \code{grow_anacata} runs a numerical simulation over time.
}
\seealso{
Other growth models: \code{\link{ode_monod}}

Other growth models: \code{\link{ode_monod}}
}
