% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeModels.R
\name{ode_anacatr}
\alias{ode_anacatr}
\alias{grow_anacatr}
\title{Regulated Anabolism vs. Catabolism Growth Model}
\usage{
ode_anacatr(time, init, parms, ...)

grow_anacatr(time, parms, ...)
}
\arguments{
\item{time}{actual time (for the ode) resp. vector of simulation time steps.}

\item{init}{named vector with the initial state of the system}

\item{parms}{parameters of the anabolism vs. catabolism growth model
\itemize{
   \item \code{y} initial biomass concentration (cell number),
   \item \code{s} initial substrate concentration
   \item \code{atp} initial ATP concentration (fraction of ATP+ADP=1)
   \item \code{phi} culture dilution rate
   \item \code{sin} substrate concentration in medium influx
   \item \code{mumax_ab} maximum growth rate (1/time),
   \item \code{mumax_cd} maximum catabolic rate (1/time),
   \item \code{mumax_m} maintenance rate concentration/time
   \item \code{n_ab} anabolic ATP stoichiometry (n ATP / C-mol substrate)
   \item \code{n_cd} catabolic ATP stoichiometry (n ATP / C-mol substrate)
   \item \code{Kaab} anabolic ATP saturation or half-maximal rate constant
   \item \code{Ksab} anabolic substrate saturation or half-maximal rate constant
   \item \code{Kacd} catabolic ATP saturation or half-maximal rate constant
   \item \code{Kscd} catabolic substrate saturation or half-maximal rate constant
   \item \code{Cc} and \code{Vc} cellular carbon content and volumes
}}

\item{\dots}{placeholder for additional parameters (for user-extended versions of this function)}
}
\value{
For \code{ode_anacatr}: matrix containing the simulation outputs.
The return value of has also class \code{deSolve}.

For \code{grow_anacatr}: vector of dependent variable (\code{y}) and
  its log-transformed values (\code{log_y}):

\itemize{
\item \code{time} time of the simulation
\item \code{y} cell concentration
}
}
\description{
Regulated Anabolism vs. Catabolism Growth Model
}
\details{
Function \code{ode_anacatr} is the system of differential equations,
whereas \code{grow_anacatr} runs a numerical simulation over time.
}
\seealso{
Other growth models: \code{\link{ode_anacat}},
  \code{\link{ode_monod}}

Other growth models: \code{\link{ode_anacat}},
  \code{\link{ode_monod}}
}
