% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeModels.R
\name{ode_monod}
\alias{ode_monod}
\alias{grow_monod}
\title{Monod's Growth Model}
\usage{
ode_monod(time, init, parms, ...)

grow_monod(time, parms, ...)
}
\arguments{
\item{time}{actual time (for the ode) resp. vector of simulation time steps.}

\item{init}{named vector with state of the system}

\item{parms}{parameters of Monod's growth model:
\itemize{
   \item \code{y} initial biomass concentration (cell number),
   \item \code{s} initial substrate concentration
   \item \code{mumax} maximum growth rate (1/time),
   \item \code{K} saturation or half-maximal growth constant
   \item \code{Y} the yield factor (fraction of substrate converted to biomass
   \item \code{phi} culture dilution rate
   \item \code{sin} substrate concentration in medium influx
}}

\item{\dots}{placeholder for additional parameters (for user-extended versions of this function)}
}
\value{
For \code{ode_monod}: matrix containing the simulation outputs.
The return value of has also class \code{deSolve}.

For \code{grow_monod}: vector of dependent variable (\code{y}) and
  its log-transformed values (\code{log_y}):

\itemize{
\item \code{time} time of the simulation
\item \code{y} cell concentration
}
}
\description{
System of two differential equations describing bacterial growth as an
autocatalytic process of substrate to biomass conversion

Simple monod growth model
}
\details{
The model is given as a system of two differential equations:

\eqn{mu = mumax * s/(s+K)}
\deqn{ds/dt = phi*(sin-s) - (mu/Y)*y}
\deqn{dy/dt = (mu-phi)*y}

Function \code{ode_monod} is the system of differential equations,
whereas \code{grow_monod} runs a numerical simulation over time.
}
\examples{

time <- seq(0, 30, length=200)
parms <- c(phi=0, sin=0, mumax=1, K=1, Y=0.5)
y0    <- c(s=1, y=0.01)
out   <- deSolve::ode(y0, time, ode_monod, parms)

plot(out)

o <- grow_monod(0:100, c(y0, parms))
plot(o)

}
\references{
Monod, J (1950). La technique de la culture continue, th\'eorie et applications. Annales de Institute Pasteur Paris 79(4), 390--410.
}
\seealso{
Other growth models: \code{\link{ode_anacatr}},
  \code{\link{ode_anacat}}

Other growth models: \code{\link{ode_anacatr}},
  \code{\link{ode_anacat}}
}
