% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R, R/interfaces.R.bak.R
\name{growthratesResults}
\alias{growthratesResults}
\alias{growthratesResults}
\title{parse results \code{\link[growthrates:growthrates]{growthrates}}}
\usage{
growthratesResults(fit)

growthratesResults(fit)
}
\arguments{
\item{fit}{growthrates object, the result of a call to
\code{\link[grofit:gcFit]{gcFit}}
or the \code{platexpress} version \code{\link{gcFit.2}}}

\item{fit}{growthrates object, the result of a call to
\code{\link[grofit:gcFit]{gcFit}}
or the \code{platexpress} version \code{\link{gcFit.2}}}
}
\description{
parses the output of \code{\link{gcFit.2}} into a table
of the main model parameters, for each well

parses the output of \code{\link{gcFit.2}} into a table
of the main model parameters, for each well
}
\seealso{
\code{\link{data2growthrates}}, \code{\link{grofitResults}}

\code{\link{data2growthrates}}, \code{\link{grofitResults}}
}
