% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{image_plate}
\alias{image_plate}
\title{Plot a heatmap of \code{platexpress} data.}
\usage{
image_plate(data, yid, wells, well.norm = FALSE, q.cut = 0.99, ylab = "",
  xid, xlab, wcol, col = gray.colors(100, start = 1, end = 0), ...)
}
\arguments{
\item{data}{\code{platexpress} data object}

\item{yid}{ID of the \code{platexpress} data to use}

\item{wells}{subset or order of wells}

\item{well.norm}{normalize each well}

\item{q.cut}{cut values higher then this quantile for color scale}

\item{ylab}{y-axis label}

\item{xid}{x-axis data ID in \code{platexpress} object}

\item{xlab}{x-axis label}

\item{wcol}{well colors (named vector) for axis labels,
eg. from plate layout map}

\item{col}{color palette to use in heatmap}

\item{...}{arguments passed to \code{\link{image}}}
}
\description{
Plot a heatmap of \code{platexpress} data.
}
