% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{interpolatePlateData}
\alias{interpolatePlateData}
\title{interpolate one dataset to common points of another
data set, e.g., fluorescence to OD}
\usage{
interpolatePlateData(data, xid, dids, n, xout)
}
\arguments{
\item{data}{the list of measurement data as provided by
\code{\link{readPlateData}}}

\item{xid}{ID of the data set which should serve as the new x-axis
all data will be interpolated to equi-spaced points along the range
of measured values}

\item{dids}{restrict interpolation to these data IDs}

\item{n}{specify the number of interpolation points, if missing the
original number of rows will be used}

\item{xout}{specify the interpolation points directly}
}
\description{
interpolate one dataset to common points of another
data set, e.g., fluorescence to OD
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
