% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{interpolatePlateTimes}
\alias{interpolatePlateTimes}
\title{interpolate all data to an average master time:
calculates average time for each measurement point
and interpolates all values to this time}
\usage{
interpolatePlateTimes(data, verb = TRUE)
}
\value{
returns a copy of the full data list with a master time and temperature added at the top level
}
\description{
interpolate all data to an average master time:
calculates average time for each measurement point
and interpolates all values to this time
}

