% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{mergeResults}
\alias{mergeResults}
\title{merge results to plate layout map}
\usage{
mergeResults(x, y, ID, by = "well", all.x = TRUE, sort = FALSE, ...)
}
\arguments{
\item{x}{a plate layout map or any data frame with a "well" column
specified by argument \code{by}}

\item{y}{results: any data frame with a "well" column
specified by argument \code{by}, see \code{\link[base:merge]{merge}}}

\item{ID}{prefix added to to the merged \code{y} data}

\item{by}{column name(s) by which to merge,}

\item{all.x}{keep all rows of \code{x}, see \code{\link[base:merge]{merge}}}

\item{sort}{sort the results on the \code{by} column? See
\code{\link[base:merge]{merge}}}

\item{...}{further arguments to \code{\link[base:merge]{merge}}}
}
\description{
a wrapper around R base function \code{\link[base:merge]{merge}}
for merging results per well into a \pkg{platexpress} layout map.
Usage is the same as for \code{\link[base:merge]{merge}} with
some defaults changed.
}
