% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parseAmounts}
\alias{parseAmounts}
\title{Parse Amount Strings}
\usage{
parseAmounts(str, sep = ":")
}
\arguments{
\item{str}{a vector of strings, providing substance:amount pairs}

\item{sep}{the separator of the substance:amount pair strings}
}
\description{
splits a vector of strings of substance:amount pairs into
a matrix with 2 columns, of substances and amounts
}
\seealso{
\code{\link{replaceAmounts}}
}
