% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\docType{package}
\name{platexpress}
\alias{platexpress}
\alias{platexpress-package}
\title{platexpress: A package for analysing microbial growth & expression.}
\description{
The platexpress package provides a quick&easy interface to 
microbial growth & gene expression data as measured in typical
microplate-readers or other growth systems.
}
\section{Platexpress functions}{

The platexpress functions
}

