% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{plotWavelength}
\alias{plotWavelength}
\title{\code{\link{plotWavelength}}:
a color selector for \code{\link{showSpectrum}}; draws a vertical line,
the wavelength, and a filled circle at a given wavelength in nm.}
\usage{
plotWavelength(x = 534, y = 1.09, ych = 0.95, cex = 5, pch = 19, ...)
}
\arguments{
\item{x}{wavelength in nm}

\item{y}{position of the text, from 0.9 to 1.1}

\item{ych}{y position of the symbol}

\item{cex}{the plot symbol size, see ?par("cex")}

\item{pch}{the plot symbol type, see ?par("pch")}

\item{...}{further arguments to ?text}
}
\description{
\code{\link{plotWavelength}}:
a color selector for \code{\link{showSpectrum}}; draws a vertical line,
the wavelength, and a filled circle at a given wavelength in nm.
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
\seealso{
\code{\link{showSpectrum}}, \code{\link{findWavelength}},
\code{\link{wavelength2RGB}}
}

