% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readBMG2Plate}
\alias{readBMG2Plate}
\title{Read BMG Optima/MARS vXXX Plate Data}
\usage{
readBMG2Plate(files, data.ids, time.format = " \%H h \%M min",
  skip = 6, sep = ";", dec = ",", verb = TRUE)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{time.format}{format of the time, e.g., "%H:%M"%S" (see
\code{strptime}),}

\item{skip}{lines to skip before parsing data}

\item{sep}{field separator used in the input tabular file}

\item{dec}{decimal number symbol (e.g., ',' if data export was with
german language settings)}

\item{verb}{print messages if true}
}
\description{
Parses date from a CSV file that can be exported
from the MARS (vXXX) analysis software of BMG plate readers. Different
measurements (eg. OD and fluorescence). Data exported from an earlier
version of MARS (vR3.01 R2) can be parsed with \code{\link{readBMGPlate}}
}
\seealso{
\code{\link{readPlateData}}, \code{\link{readBMGPlate}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
