% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readExperiment}
\alias{readExperiment}
\title{High-level interface of \code{\link{platexpress}}}
\usage{
readExperiment(
  files,
  type,
  dsep,
  time.range = c("full"),
  time.conversion = 1/3600,
  layout,
  sep = ";",
  fsep = ";",
  asep = ":",
  blank.id = "blank",
  fields,
  afields,
  skip.wells,
  blank = FALSE,
  blank.data,
  base = 0,
  group1,
  group2,
  group2.color,
  ...
)
}
\arguments{
\item{files}{list of one or more data files}

\item{type}{pre-defined formats, as exported from platereaders; currently
for BMG Optima and Mars v3.01 R2, ('BMG'), BMG Clariostar and Mars
vXXX ('BMG2') and Biotek Synergy Mx ('Synergy'), BioLector ('BioLector'), BioLector Pro ('BioLectorPro');  TODO: define export
protocols!}

\item{dsep}{\code{sep} parameter for data file, see
\code{\link{readPlateData}}}

\item{time.range}{"common" requires that all data must have
the same number of time-points, "full" interpolates
the maximal available time range, using the mean
time step of the mean of data-specific time vectors}

\item{time.conversion}{conversion factor for the plate time, e.g., 1/3600
to convert from hours to seconds}

\item{layout}{the text file containing the plate layout information, see
argument \code{file} in \code{\link{readPlateMap}}}

\item{sep}{column separator, as in \code{\link[utils:read.table]{read.table}}}

\item{fsep}{within-field separator, separating the well-specific descriptors within
well fields}

\item{asep}{a separator for substance:amount pairs, eg. to indicate amount
of an inducer or a nutrient, can only be used together with
argument \code{afields}, eg. use \code{asep=":", afields="inducer"}}

\item{blank.id}{keyword that indicates blank wells. Blank wells can be
combined with other well descriptors for separate blanking}

\item{fields}{names for the field descriptors}

\item{afields}{field names which hold substance:amount pair information,
see argument \code{asep}}

\item{skip.wells}{skip these wells from all analyses, see argument
\code{skip} in \code{\link{skipWells}}}

\item{blank}{call \code{\link{correctBlanks}} to blank data}

\item{blank.data}{subset of data to blank, see argument \code{yids} in
function \code{\link{correctBlanks}}}

\item{base}{optional minimal value; all values will be raised by
the same amount using the function \code{\link{adjustBase}} via
function \code{\link{correctBlanks}}}

\item{group1}{plate layout column to be used for coarse grouping, via
function \code{\link{getGroups}}, eg. "strains"}

\item{group2}{plate layout column to be used for fine grouping, via
function \code{\link{getGroups}}, eg. "inducer amount")}

\item{group2.color}{numeric plate layout column to be used for group2
coloring, via function \code{\link{groupColors}}, eg. "amount"}

\item{...}{further parameters to plate-reader specific parsing functions}
}
\description{
This high-level wrapper allows to load a platereader experiment
and its layout file, do blank corrections, and replicate groupings.
}
\details{
Calls, in this order the  \code{\link{platexpress}} functions
 \code{\link{readPlateData}},  \code{\link{readPlateMap}};
 optionally: \code{\link{skipWells}}, and \code{\link{correctBlanks}},
and \code{\link{getGroups}}. Please see the help files of these
functions for details on parameters.
}
