% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readPlateMap.old}
\alias{readPlateMap.old}
\title{Read Plate Layout Map}
\usage{
readPlateMap.old(file, sep = "\\t", fsep = "\\n", fields, asep, afields,
  blank.id = "blank", nrows = 8, formatted = FALSE, header = TRUE)
}
\arguments{
\item{file}{text file containing the plate layout.}

\item{sep}{column separator, as in \code{\link[utils:read.table]{read.table}}}

\item{fsep}{within-field separator, separating the well-specific descriptors within
well fields}

\item{fields}{names for the field descriptors}

\item{asep}{a separator for substance:amount pairs, eg. to indicate amount
of an inducer or a nutrient, can only be used together with
argument \code{afields}, eg. use \code{asep=":", afields="inducer"}}

\item{afields}{field names which hold substance:amount pair information,
see argument \code{asep}}

\item{blank.id}{keyword that indicates blank wells. Blank wells can be
combined with other well descriptors for separate blanking}

\item{nrows}{number of rows to expect, defaults to 8 for rows A to H in
a typical 96 well plate; TODO: get rid of this and instead check for
rownames?}

\item{formatted}{indicates whether the file is already in the required
format; all other paramaters but 'sep' will be ignored}

\item{header}{logical argument indicating the presence/absence of a header
row in the layout file}
}
\value{
a table of well content descriptors, where the first column 'wells'
        maps the plate map to the data files.
}
\description{
Parses a plate design file in CSV format. Rows and
columns should be named as in the corresponding plate reader data files.
}
\examples{
plate.file <- system.file("extdata", "AP12_layout.csv", package = "platexpress")
plate <- readPlateMap(file=plate.file, blank.id="blank",fsep="\\n", fields=c("strain","samples"))
}
\seealso{
\code{\link{readPlateData}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
