% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{readPlateMap}
\alias{readPlateMap}
\title{\code{\link{readPlateMap}} parses a plate design file in CSV. Rows and 
columns should be named as in the corresponding datafiles.}
\usage{
readPlateMap(file, sep = "\\t", fsep = "\\n", blank.id = "blank", fields,
  formatted = FALSE)
}
\arguments{
\item{file}{text file containing the plate layout.}

\item{sep}{column separator, as in read.table}

\item{fsep}{within-field separator, separating the well-specific descriptors}

\item{blank.id}{keyword that indicates blank wells. Blank wells can be
combined with other well descriptors for separate blanking}

\item{fields}{names for the field descriptors}

\item{formatted}{indicates whether the file is already in the required
format; all other paramaters but 'sep' will be ignored}
}
\value{
a table of well content descriptors, where the first column 'wells'
        maps the plate map to the data files.
}
\description{
\code{\link{readPlateMap}} parses a plate design file in CSV. Rows and 
columns should be named as in the corresponding datafiles.
}
\examples{
plate.file <- system.file("extdata", "AP12_layout.csv", package = "platexpress")
plate <- readPlateMap(file=plate.file, blank.id="blank",fsep="\\n", fields=c("strain","samples"))
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
\seealso{
\code{\link{readPlateData}}
}

