% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readSimplePlate}
\alias{readSimplePlate}
\title{Read in simple platereader data with only one data type, and time
in the first column}
\usage{
readSimplePlate(files, data.ids, skip, time.format = "\%M:\%S", verb = TRUE)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{skip}{lines to skip before parsing data}

\item{verb}{print messages if true}

\item{sep}{field separator used in the input tabular file}

\item{dec}{decimal number symbol (e.g., ',' if data export was with
german language settings)}
}
\description{
Read in simple platereader data with only one data type, and time
in the first column
}

