% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readSynergyPlate}
\alias{readSynergyPlate}
\title{Read Synergy Mx Plate Data}
\usage{
readSynergyPlate(files, data.ids, skip, sep = ";", dec = ".",
  skiplastcol = FALSE, time.format = "\%H:\%M:\%S", verb = TRUE)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{skip}{lines to skip before parsing data}

\item{sep}{field separator used in the input tabular file}

\item{dec}{decimal operator used in data file (e.g., ',' if data export
was with german language settings)}

\item{skiplastcol}{the last column is often empty, set to TRUE if
this is the case for the current data set}

\item{time.format}{format of the time, e.g., "%H:%M"%S" (see
\code{strptime}), or "numeric" if the time is provided in normal numbers}

\item{verb}{print messages if true}
}
\description{
Parses date exported from the Excel file that can be exported
from the Synergy Mx platereader. A parameter that often changes
is \code{skip}, the number of lines before the data starts,
here before the ID of a measurement in the first column, eg.
"600" for OD measurement at 600 nm.
}
\seealso{
\code{\link{readPlateData}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
