% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{replaceAmounts}
\alias{replaceAmounts}
\title{Numeric Amounts}
\usage{
replaceAmounts(map, col, sep = ":", replace = TRUE)
}
\arguments{
\item{map}{the plate map, see  \code{\link{readPlateMap}}}

\item{col}{the column ID or index to be split}

\item{sep}{the separator of the substance:amount pair strings}

\item{replace}{if TRUE the original columns in \code{col} will be removed}
}
\description{
replaces amount strings of the form substance:amount in columns of
plate layout maps by two columns, separating the substance and the amount;
wrapper for \code{\link{parseAmounts}}
}
