% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{rmData}
\alias{rmData}
\title{\code{\link{rmData}} : remove a data set}
\usage{
rmData(data, ID)
}
\arguments{
\item{data}{the current platexpress data set}

\item{ID}{a vector of IDs of the data to be removed}

\item{dat}{the new data, must be a matrix akin to other data in the set}
}
\description{
\code{\link{rmData}} : remove a data set
}
\seealso{
\code{\link{addData}}
}

