% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{shiftData}
\alias{shiftData}
\title{Shift Data}
\usage{
shiftData(data, lag, mid)
}
\arguments{
\item{data}{\code{\link{platexpress}} data set}

\item{lag}{a named vector providing the lag-phase to be removed; the names
correspond to the wells in data}

\item{mid}{the x-axis to be used, defaults to the first available
(usually "Time")}
}
\description{
Shifts the x-axis, eg. by calculated lag-phases, to align growth curves
}
