% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{viewGroups}
\alias{viewGroups}
\title{plot grouped wells as summary plots, incl. confidence intervals and means}
\usage{
viewGroups(data, groups, mids = c(time = "Time", temp = "Temperature"), xid,
  xscale = FALSE, xlim, dids, pcols, yscale = TRUE, ylims, ylim, log = "",
  legpos = "topleft", lwd.orig = 0.5, mai = c(0.5, 0, 0, 0),
  mgp = c(1.5, 0.5, 0), nrow = 1, xaxis = TRUE, yaxis = TRUE)
}
\arguments{
\item{data}{the list of measurement data as provided by \code{\link{readPlateData}}}

\item{groups}{a list of well grouped wells, as produced by \code{\link{getGroups}}}

\item{mids}{vector of named strings, indicating the IDs of the master
time and temperature vectors in the data}

\item{xid}{ID of a data-set in the input data that can be used as x-axis
instead of the default Time vector}

\item{xscale}{use a global range for the x-axes; only relevant if xid specifies a subset of the data as x-axis}

\item{xlim}{plot range of the x-axis}

\item{dids}{IDs of the data to be plotted; if missing, all data will
be plotted}

\item{pcols}{a named list of RGB colors to be used the plotted data types; the color vector must have names according to the data IDs}

\item{yscale}{if TRUE (default) global y-axis limits will be calculated from
all plotted wells; if FALSE each well be locally scaled}

\item{ylims}{a named list of y-axis ranges pairs for each data ID}

\item{ylim}{one y-axis limit range that will be used for all plotted data}

\item{log}{plot logarithmic axis, use equivalent to normal plot 'log', i.e.,
log="y" for a log y-axis, log="x" for x-axis and log="yx" for both axes}

\item{legpos}{position of the well IDs on the plots}

\item{lwd.orig}{line-width of the original single data, set to 0 to supress plotting of all original data}

\item{mai}{set the outer margins around plot areas, see ?par}

\item{mgp}{set the position of axis title, tick marks and tick lengths}

\item{xaxis}{plot x-axis if TRUE}

\item{yaxis}{plot y-axis if TRUE}

\item{nrows}{number of plot rows}
}
\description{
plot grouped wells as summary plots, incl. confidence intervals and means
}
\examples{
data(ap12)
groups <- getGroups(plate=ap12plate, by=c("strain"))
vg <- viewGroups(ap12data,groups=groups,lwd.orig=0.1,nrow=1)
}
\seealso{
\code{\link{viewPlate}}, \code{\link{getGroups}}, \code{\link{readPlateMap}}
}

