% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{viewMap}
\alias{viewMap}
\title{Plot a representation of microreader "plate"}
\usage{
viewMap(map, nrow = 8, ncol = 12, color = "color", text = "amount",
  title = "plate layout", add = FALSE)
}
\arguments{
\item{map}{the plate map, see  \code{\link{readPlateMap}}}

\item{nrow}{number of rows of the plate, with letter index from top to bottom}

\item{ncol}{number of columns of the plate}

\item{color}{color column in the \code{map}}

\item{text}{text column in the \code{map}}

\item{title}{plot title (argument \code{main} to plot)}

\item{add}{add to an existing plot}

\item{...}{arguments passed to \code{\link{image}}}
}
\description{
Plot a representation of microreader "plate"
}
