% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{viewSpectrum}
\alias{viewSpectrum}
\title{\code{\link{viewSpectrum}}:
shows the color spectrum of visible light along wavelengths in nm.
See \code{\link{wavelength2RGB}} for details.}
\usage{
viewSpectrum(wavelengths = 380:780, alpha = 99, pch = 19, cex = 3,
  ylab = "approximate color", main = "use findWavelength(353)",
  xlab = "wavelength, nm", ...)
}
\arguments{
\item{wavelengths}{vector of wavelengths to be plotted}

\item{alpha}{the alpha factor (opacity) for plot symbols, min=1, max=255}

\item{pch}{the plot symbol type, see ?par("pch")}

\item{cex}{the plot symbol size, see ?par("cex")}

\item{ylab}{ylab, see ?plot}

\item{main}{main, see ?plot}

\item{xlab}{xlab, see ?plot}

\item{...}{further arguments to plot}
}
\description{
\code{\link{viewSpectrum}}:
shows the color spectrum of visible light along wavelengths in nm.
See \code{\link{wavelength2RGB}} for details.
}
\details{
Plots the color spectrum over the selected wavelengths in nm.
The utility functions  \code{\link{plotWavelength}} and
\code{\link{findWavelength}} can be used to plot and interactively (click
on plot) find wavelengths. The function \code{\link{wavelength2RGB}} can
then be used to convert this wavelength to RGB colors.
}
\seealso{
\code{\link{wavelength2RGB}}, \code{\link{findWavelength}},
\code{\link{plotWavelength}}
}

