% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{wavelength2RGB}
\alias{wavelength2RGB}
\title{Convert Wavelength to RGB Colors}
\usage{
wavelength2RGB(wavelength)
}
\arguments{
\item{wavelength}{the wavelength in nm to convert to RGB}
}
\description{
Converts wavelength in nm (visible light: 380:780 nm) to RGB.
NOTE that this is not a fully correct spectrum, see Details.
}
\details{
implemented following the java code posted at
http://stackoverflow.com/questions/1472514/convert-light-frequency-to-rgb .
Also see http://www.fourmilab.ch/documents/specrend/ for original code and
why not all wavelengths can be converted to RGB.
}
\examples{
wavelengths <- seq(380,780,1)
cols <- sapply(wavelengths, wavelength2RGB)
bars <- rep(1,length(wavelengths)); names(bars) <- wavelengths
barplot(bars,border=cols,col=cols,las=2)
}
\seealso{
\code{\link{showSpectrum}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
