% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{addModel}
\alias{addModel}
\title{Add \pkg{grofit}/\pkg{growthrates} fits to plate data object}
\usage{
addModel(data, fit, ID = "model", ...)
}
\arguments{
\item{data}{a platexpress data set, see \code{\link{readPlateData}}}

\item{fit}{result object from calls to batch model fitting functions
from \pkg{grofit} or \pkg{growthrates}, ie. result of a call to
\code{\link{gcFit.2}} or any of \pkg{growthrates} batch functions (`all_')}

\item{ID}{ID for the new data set}

\item{...}{arguments to \code{\link{addData}} (eg. \code{col} for
color selection)}
}
\description{
Add \pkg{grofit}/\pkg{growthrates} fits to plate data object
}
