% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{addModel}
\alias{addModel}
\title{add grofit models to plate data object}
\usage{
addModel(data, fit, ID = "model", ...)
}
\arguments{
\item{data}{a platexpress data set, see \code{\link{readPlateData}}}

\item{fit}{grofit object, the result of a call to \code{\link{gcFit.2}}}

\item{ID}{ID for the new data set}

\item{...}{arguments to \code{\link{addData}}}
}
\description{
add grofit models to plate data object
}
