% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{addModel}
\alias{addModel}
\title{Add fits from various growth model interfaces to plate data object.}
\usage{
addModel(data, fit, ID = "model", ...)
}
\arguments{
\item{data}{a platexpress data set, see \code{\link{readPlateData}}}

\item{fit}{result object from calls to batch model fitting functions
\code{\link{gcFit.2}}, any of \pkg{growthrates} batch functions (`all_'),
or \code{platexpress} interfaces \code{\link{segmented_plate}}, and
\code{\link{dpseg_plate}}.}

\item{ID}{ID for the new data set}

\item{...}{arguments to \code{\link{addData}} (eg. \code{col} for
color selection)}
}
\description{
TODO: depending on class of `fit`.
}
