% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{adjustBase}
\alias{adjustBase}
\title{\code{\link{adjustBase}} adjust data to a minimal base}
\usage{
adjustBase(data, dids, base = 0)
}
\arguments{
\item{dids}{vector of ID strings for which base correction should be
executed}

\item{base}{the new minimum for the data, default is 0, but it could
e.g. be the OD used for inoculation}

\item{by}{TODO: choose specific groups via plate-designs}
}
\value{
Returns `data' where all data sets or only those selected by option
dids where raised to a minimum level in
}
\description{
\code{\link{adjustBase}} adjust data to a minimal base
}
\details{
Adjusts data to a new mininum, this is useful for adjustment
of negative values after blank corrections
}

