% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{amountColors}
\alias{amountColors}
\title{Add Amount Colors}
\usage{
amountColors(map, substance = "substance", amount = "amount",
  colf = colorRamps::matlab.like2)
}
\arguments{
\item{map}{the plate map, see  \code{\link{readPlateMap}}}

\item{substance}{name of the substance column in \code{map}}

\item{amount}{name of the amount column in \code{map}}

\item{colf}{color palette function}
}
\description{
simply add a color palette along the range of added amounts
of a given substance; where substance and amount columns
obtained from from \code{\link{replaceAmounts}}, or options\\code{asep}
and \code{afield} in \code{\link{readPlateMap}}.
}
\seealso{
\code{\link{readPlateMap}}, \code{\link{replaceAmounts}}
}
