% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\docType{data}
\name{ap12data}
\alias{ap12data}
\title{ap12data. data by Dennis Dienst and Alice Pawloski, incl. the
plate reader measurements of E.coli growth, expressing a fluorescent
proteins, in a Synergy Mx platereader; the corresponding plate layout
map is in \code{\link{ap12plate}}.}
\description{
\itemize{
  \item Data:
  \item Time: the interpolated 'master' time
  \item Temperature: the temperature time-course
  \item Data matrix '600': well absorbance at 600 nm, i.e., the OD,
  \item Data matrix 'YFP_50:500,535': the YFP fluorescence measured by excitation at 500 nm and emission at 535 nm 
}
}

