% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxData}
\alias{boxData}
\title{Box-Plots of Data Ranges}
\usage{
boxData(data, rng, groups, xid, yid = "OD", interpolate = TRUE,
  plot = TRUE, type = "box", etype = "ci")
}
\arguments{
\item{data}{\code{\link{platexpress}} data, see \code{\link{readPlateData}}}

\item{rng}{x-axis range or a single point, for the latter the closest
point will be selected or, if option \code{interpolate=TRUE} the
value at \code{rng} will be interpolated using R base function
\code{\link[stats:approxfun]{approx}}}

\item{groups}{a grouping of wells, see \code{\link{getGroups}}}

\item{xid}{the x-axis ID, if multiple x-axes are present}

\item{yid}{the y-axis data to be grouped}

\item{interpolate}{interpolate data at \code{rng} (if a single value,
see argument \code{rng})}

\item{plot}{if TRUE a box-plot or bar-plot will be plotted}

\item{type}{either "box" (default) or "bar" for box-plot or bar-plot}

\item{etype}{type of statistics to be used for error bars in the bar-plot,
either "ci" (default) for the 95%-confidence interval or "se" for
the standard error}
}
\value{
Returns an annotated \code{data.frame} of the values, with well
and group IDs in the first two columns. The values in the third column are
raw values (if argument \code{rng} was a single point
on the x-axis) or mean values (if argument\code{rng} was a range).
}
\description{
returns data for group of wells in a given range of the x-axis;
if only one value is given, the values closest to this x-axis points
are returned; see "Value" for details.
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
