% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{correctBlanks}
\alias{correctBlanks}
\title{\code{\link{correctBlanks}} correct for blanks}
\usage{
correctBlanks(data, plate, type = "ci95", by, dids, mid = "Time", max.mid,
  mbins = 1)
}
\arguments{
\item{data}{the data list to be blank-corrected}

\item{plate}{the plate layout where column "blanks" indicates which wells
are to be treated as blanks}

\item{by}{a list of column IDs of the plate layout; separate blank
correction will be attempted for groups in these columns; each group
must have at least one specified blank associated}

\item{dids}{IDs of the data which should be blank-corrected, all will be
blanked if missing}
}
\description{
\code{\link{correctBlanks}} correct for blanks
}
\examples{
data(ap12)
data <- correctBlanks(data=ap12data, plate=ap12plate, by="strain")
}

