% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{cutData}
\alias{cutData}
\title{\code{\link{cutData}} : cut data in a range of the x-axis}
\usage{
cutData(data, rng, mid)
}
\arguments{
\item{data}{\code{\link{platexpress}} data, see \code{link{readPlateData}}}

\item{rng}{a single value or a data range}
}
\description{
\code{\link{cutData}} : cut data in a range of the x-axis
}
\details{
Cuts the passed \code{\link{platexpress}} data to ranges of
of the x-axis (time or other, see \code{data$mids}). If paramter \code{rng}
is a single value, data for the closest x value will be returned
if rng is a single value
}

