% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{data2grofit}
\alias{data2grofit}
\title{interface to package \code{\link[grofit:grofit]{grofit}}}
\usage{
data2grofit(data, yid, min.time, max.time, wells, plate, eid, dose)
}
\arguments{
\item{data}{a platexpress data set, see \code{\link{readPlateData}}}

\item{yid}{data ID of the data to be converted for grofit, from \code{data$dataIDs}}

\item{min.time}{minimal time of the data to be used}

\item{max.time}{maximal time of the data to be used}

\item{wells}{column IDs of the data set to use, if missing all wells
are used}

\item{plate}{plate layout map, see \code{\link{readPlateMap}}, columns
of this map can be converted to \code{\link[grofit:grofit]{grofit}} data
annotation}

\item{eid}{column IDs in the plate layout map to be used for
\code{\link[grofit:grofit]{grofit}} data annotation; if missing but
\code{plate} is present, the columns 2 and 3 are used}

\item{dose}{vector of doses in each well, used as the third column of
 \code{\link[grofit:grofit]{grofit}} data annotation, where it can be used for
dose-response calculations}
}
\description{
\code{\link{data2grofit}} : converts \code{\link{platexpress}} data to
\code{\link[grofit:grofit]{grofit}} data format
}
\details{
Returns a simple list with two entries \code{time} and \code{data},
as required for \code{\link[grofit:grofit]{grofit}}.
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
