% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{data2growthrates}
\alias{data2growthrates}
\title{interface to package \pkg{growthrates}}
\usage{
data2growthrates(data, yid)
}
\arguments{
\item{data}{platexpress data set, see \code{\link{readPlateData}}}

\item{yid}{data ID of the data to be converted for growthrates, from
\code{data$dataIDs}}
}
\description{
converts \code{platexpress} data for
use with \pkg{growthrates}
}
\seealso{
\code{\link{growthratesResults}}, \code{\link{data2grofit}}
}
