% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{fitCellGrowths.2}
\alias{fitCellGrowths.2}
\title{batch wrapper for fitCellGrowth
TODO: align in/out with gcFit.2 or optionally cellGrowth::fitCellGrowths}
\usage{
fitCellGrowths.2(data, did, mid, wells, ...)
}
\description{
batch wrapper for fitCellGrowth
TODO: align in/out with gcFit.2 or optionally cellGrowth::fitCellGrowths
}

