% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{getData}
\alias{getData}
\title{\code{\link{getData}} : get a specific data set, returns a data matrix}
\usage{
getData(data, ID, type = "data")
}
\arguments{
\item{data}{the current platexpress data set}

\item{ID}{the ID of the data to be obtained}

\item{type}{the type of the data to be returned (default "data"),
the data list also contains original values for some of the processing
steps}
}
\description{
\code{\link{getData}} : get a specific data set, returns a data matrix
}
