% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{getData}
\alias{getData}
\title{Get Data}
\usage{
getData(data, ID, type = "data", xrng, xid, verb = TRUE)
}
\arguments{
\item{data}{the current platexpress data set}

\item{ID}{the ID of the data to be obtained}

\item{type}{the type of the data to be returned (default "data"),
the data list also contains original values for some of the processing
steps}

\item{xrng}{x-axis range or single point; similar to cutData, but
use interpolation for single values!}

\item{xid}{ID of the x-axis data to use}

\item{verb}{print messages}
}
\description{
Returns a specific data set as data matrix
}
