% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{getGroups}
\alias{getGroups}
\title{Get Groups of Replicates}
\usage{
getGroups(plate, by = "medium", order = FALSE, verb = TRUE)
}
\arguments{
\item{plate}{the plate layout map, see \code{\link{readPlateMap}}}

\item{by}{a list of column IDs of the plate layout}

\item{order}{if TRUE groups will be alphabetically ordered}

\item{verb}{if TRUE report messages are more detailed}
}
\value{
Returns a list of well IDs for the identified grouping. This list
can be used, e.g., in viewGroups(data,groups) or \code{link{groupStats}}
to summarize data for these groups.
}
\description{
groups wells by experiment annotations in the plate layout map
(selected by option \code{by}), and returns a list of well IDs 
that are all replicates for these groups.
}
\details{
Calculates the distinct groups from the plate layout by the selected
experimental parameters.
}
\seealso{
\code{\link{readPlateMap}}, \code{\link{viewGroups}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
