% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R, R/interfaces.R.bak.R
\name{grofit.2.control}
\alias{grofit.2.control}
\alias{grofit.2.control}
\title{wrapper of grofit function \code{\link[grofit:grofit.control]{grofit.control}}
which adds the new \code{plot} switch used in \code{\link{gcFit.2}}}
\usage{
grofit.2.control(interactive = FALSE, plot = TRUE, ...)

grofit.2.control(interactive = FALSE, plot = TRUE, ...)
}
\arguments{
\item{interactive}{set to TRUE for interactive growth curve fitting}

\item{plot}{set to TRUE for plots even without \code{interactive} use}

\item{...}{parameters passed on to
\code{\link[grofit:grofit.control]{grofit.control}}}

\item{...}{parameters passed on to
\code{\link[grofit:grofit.control]{grofit.control}}}

\item{interactive}{set to TRUE for interactive growth curve fitting}

\item{plot}{set to TRUE for plots even without \code{interactive} use}
}
\description{
wrapper of grofit function \code{\link[grofit:grofit.control]{grofit.control}}
which adds the new \code{plot} switch used in \code{\link{gcFit.2}}

wrapper of grofit function \code{\link[grofit:grofit.control]{grofit.control}}
which adds the new \code{plot} switch used in \code{\link{gcFit.2}}
}
\seealso{
\code{\link{gcFit.2}}

\code{\link{gcFit.2}}
}
