% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{grofitGetParameters}
\alias{grofitGetParameters}
\title{Convenience function to quickly extract growth parameters
from a \code{\link[grofit:grofit]{grofit}} results list.}
\usage{
grofitGetParameters(fit, p = c("TestId", "AddId", "lambda.model", "mu.model",
  "A.model", "used.model"))
}
\arguments{
\item{fit}{grofit object, the result of a call to 
\code{\link[grofit:gcFit]{gcFit}}
or the \code{platexpress} version \code{\link{gcFit.2}}}

\item{p}{list of parameters to obtain from the table \code{fits$gcTable}}
}
\description{
Convenience function to quickly extract growth parameters
from a \code{\link[grofit:grofit]{grofit}} results list.
}
