% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{grofitResults}
\alias{grofitResults}
\title{parse grofit results}
\usage{
grofitResults(fit)
}
\arguments{
\item{fit}{grofit object, the result of a call to \code{\link{gcFit.2}}}
}
\description{
parses the output of \code{\link{gcFit.2}} into a table
of the main model parameters, for each well
}
