% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{groupColors}
\alias{groupColors}
\title{get colors for all wells in a group}
\usage{
groupColors(plate, group, color = "color")
}
\arguments{
\item{plate}{the plate layout map, see \code{\link{readPlateMap}},
for which a coloring scheme has been specified, eg. by
\code{\link{amountColors}}}

\item{group}{a well grouping, see \code{\link{getGroups}}}

\item{color}{the name of the column in \code{plate} providing colors}
}
\description{
returns a named vector of colors for a grouping if all
colors in that group are unique
}
