% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{groupStats}
\alias{groupStats}
\title{calculates simple statistics for groups as plotted in
\code{\link{viewGroups}}}
\usage{
groupStats(data, groups, dids)
}
\arguments{
\item{data}{\code{\link{platexpress}} data, see \code{\link{readPlateData}}}

\item{groups}{a list of well grouped wells, as produced by
\code{\link{getGroups}}}

\item{dids}{data IDs for which statistics should be reported,
if missing stats for all data will be reported}
}
\value{
Returns a data structure similar to the input data, but
where actual data is replaced  statistics over grouped wells.
}
\description{
calculates simple statistics for groups as plotted in
\code{\link{viewGroups}}
}
\details{
Calculates the simple statistics over grouped wells
(means, 95% confidence intervals, stdandard errors) along the x-axis
(usually time).
}
\seealso{
\code{\link{readPlateMap}}, \code{\link{viewGroups}}
}

