% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{growthratesResults}
\alias{growthratesResults}
\title{parse fitted parameters from package \pkg{growthrates}}
\usage{
growthratesResults(fit, scale.richards = TRUE)
}
\arguments{
\item{fit}{growthrates object, the result of a call to fit functions
from package \code{\link[growthrates]{growthrates}}, such as
\code{\link[growthrates:all_easylinear]{all_easylinear}}}

\item{scale.richards}{if the \code{beta} parameter from Richard's model is present,
multiply growthrate \code{mu} with \code{beta} (original stored as \code{mumax})}
}
\description{
parses the output of \code{\link{gcFit.2}} into a table
of the main model parameters, for each well
}
\seealso{
\code{\link{data2growthrates}}, \code{\link{grofitResults}}
}
