% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{interpolatePlateData}
\alias{interpolatePlateData}
\title{Interpolate Plate Data}
\usage{
interpolatePlateData(data, xid, yids, n, xout)
}
\arguments{
\item{data}{the list of measurement data as provided by
\code{\link{readPlateData}}}

\item{xid}{ID of the data set which should serve as the new x-axis
all data will be interpolated to equi-spaced points along the range
of measured values}

\item{yids}{restrict interpolation to these data IDs}

\item{n}{specify the number of interpolation points, if missing the
original number of rows will be used}

\item{xout}{specify the interpolation points directly}
}
\description{
interpolate one dataset to common points of another
data set. This is used to change the x-axis of a data set, e.g., 
the measured OD values can become the new x-axis and all fluorescence 
values will be interpolated to common OD vlues, using the 
R base function \code{\link[stats:spline]{spline}}, the same way
as in \code{\link{interpolatePlateTimes}}.
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
