% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\docType{package}
\name{platexpress}
\alias{platexpress}
\alias{platexpress-package}
\title{platexpress: quick inspection and analysis of microbial growth & expression.}
\description{
The platexpress package provides a quick & easy interface to 
inspect microbial growth & gene expression data as measured in typical
platereaders ("96-well plates") or other parallel growth systems.
Interfaces to existing packages allow growth model fitting 
(currently only \code{\link[grofit:grofit]{grofit}}, and
partially/untested \code{\link[cellGrowth:fitCellGrowth]{cellGrowth}}.)
}
\section{Dependencies}{
 The package uses mostly functionality from R base,
(graphics, grDevices, stats) but more functionality is available when
\code{\link[grofit:grofit]{grofit}} is installed.
}

\author{
Rainer Machne
}
