% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\docType{package}
\name{platexpress}
\alias{platexpress}
\alias{platexpress-package}
\title{platexpress: A package for analysing microbial growth & expression.}
\description{
The platexpress package provides a quick&easy interface to 
microbial growth & gene expression data as measured in typical
microplate-readers or other parallel growth systems.
}
\section{Dependencies}{
 The package uses mostly functionality from R base,
(graphics, grDevices, stats) but more functionality is available when
\code{\link[grofit:grofit]{grofit}} is installed.
}

\author{
Rainer Machne
}
