% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{plotdev}
\alias{plotdev}
\title{plots to png, eps or pdf, taking the same arguments}
\usage{
plotdev(file = "test", type = "png", width = 5, height = 5, res = 100)
}
\arguments{
\item{file}{the name of the file without the extension (.pdf)}

\item{type}{type of the figure, either png, pdf or eps}

\item{width}{the figure width in inches}

\item{height}{the figure height in inches}

\item{res}{the figure resolution in ppi (pixels per inch), only used}
}
\description{
plots to png, eps or pdf, taking the same arguments
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
png
}

