% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readBMGPlate}
\alias{readBMGPlate}
\title{Read BMG Optima/MARS-exported files}
\usage{
readBMGPlate(files, data.ids, skip, sep = ";", dec = ".", verb = TRUE)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{skip}{lines to skip before parsing data}

\item{sep}{field separator used in the input tabular file}

\item{dec}{decimal number symbol (e.g., ',' if data export was with
german language settings)}

\item{verb}{print messages if true}
}
\description{
Read BMG Optima/MARS-exported files
}
\seealso{
\code{\link{readPlateData}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
