% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{readBMGPlate}
\alias{readBMGPlate}
\title{Read BMG Optima/MARS-exported files}
\usage{
readBMGPlate(files, data.ids, skip = 5, sep = ";", dec = ".",
  verb = TRUE)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{skip}{lines to skip from the data file}

\item{sep}{column separator, as in read.table}

\item{dec}{decimal operator used in the data file}

\item{verb}{print messages if true}
}
\description{
Read BMG Optima/MARS-exported files
}
\seealso{
\code{\link{readPlateData}}
}

