% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readBMGPlate}
\alias{readBMGPlate}
\title{Read BMG Optima/MARS vR3.01 R2 Plate Data}
\usage{
readBMGPlate(files, data.ids, skip, sep = ";", dec = ".",
  verb = TRUE)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{skip}{lines to skip before parsing data; if missing it will
be set to 5}

\item{sep}{field separator used in the input tabular file}

\item{dec}{decimal number symbol (e.g., ',' if data export was with
german language settings)}

\item{verb}{print messages if true}
}
\description{
Parses date from a CSV file that can be exported
from the MARS (vR3.01 R2) analysis software of BMG plate readers. Different
measurements (eg. OD and fluorescence) are exported separately
and should all be liste in parameter \code{files}. Later
versions of MARS (vXXX) have a simpler file format, and can be
parsed with \code{\link{readBMG2Plate}}
}
\seealso{
\code{\link{readPlateData}}, \code{\link{readBMG2Plate}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
