% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readSimplePlate}
\alias{readSimplePlate}
\title{Read Simple Plate Data Tables}
\usage{
readSimplePlate(files, data.ids, nrow, time.format = "\%M:\%S",
  verb = TRUE, ...)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{nrow}{organize data into a matrix with this number of rows}

\item{time.format}{format of the time, see \code{\link[base:strptime]{strptime}}); only used if time column is not numeric!}

\item{verb}{print messages if true}

\item{...}{arguments passed to \code{\link{read.csv}}}
}
\description{
Read in simple platereader data with only one data type, time
in the first column and data for all wells in the following columns,
where the column header must correspond to wells
in \code{\link{readPlateMap}}.
}
