% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readSimplePlate}
\alias{readSimplePlate}
\title{Read Simple Plate Data Tables}
\usage{
readSimplePlate(files, data.ids, skip, sep = "\\t", time.format = "\%M:\%S",
  verb = TRUE)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{skip}{lines to skip before parsing data}

\item{sep}{field separator used in the input tabular file}

\item{time.format}{format of the time, e.g., "%H:%M"%S" (see
\code{strptime}), or "numeric" if the time is provided in normal numbers}

\item{verb}{print messages if true}
}
\description{
Read in simple platereader data with only one data type, time
in the first column and data for all wells in the following columns, 
where the column header must correspond to wells
in \code{\link{readPlateMap}}.
}
