% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{readSynergyPlate}
\alias{readSynergyPlate}
\title{Read Synergy Mx-exported files}
\usage{
readSynergyPlate(files, data.ids, skip = 58, sep = ";", dec = ".",
  skiplastcol = FALSE, time.format = "numeric", verb = TRUE)
}
\arguments{
\item{files}{list of one or more data files}

\item{data.ids}{an optional sub-selection of data types in the input file,
as a list of strings}

\item{skip}{lines to skip from the data file}

\item{sep}{column separator, as in read.table}

\item{dec}{decimal operator used in the data file}

\item{skiplastcol}{the last column is often empty, set to TRUE if
this is the case for the current data set}

\item{time.format}{format of the time, e.g., "%H:%M"%S", see
\code{strptime}}

\item{verb}{print messages if true}
}
\description{
Read Synergy Mx-exported files
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
\seealso{
\code{\link{readPlateData}}
}

