% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{rmData}
\alias{rmData}
\title{\code{\link{rmData}} : remove a data set}
\usage{
rmData(data, ID)
}
\arguments{
\item{data}{a \code{\link{platexpress}} data set}

\item{ID}{a vector of IDs of the data to be removed}
}
\value{
Returns the new \\code{\link{platexpress}} data, with 
data set <\code{ID}> removed
}
\description{
\code{\link{rmData}} : remove a data set
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
\seealso{
\code{\link{addData}}
}

