% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{shiftData}
\alias{shiftData}
\title{\code{\link{shiftData}} : shift x-axis by a lag-phase to align growth curves}
\usage{
shiftData(data, lag, dids, mid)
}
\arguments{
\item{lag}{a named vector given the lag-phase to be removed; the names correspond to the wells in data}
}
\description{
\code{\link{shiftData}} : shift x-axis by a lag-phase to align growth curves
}

