% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{skipWells}
\alias{skipWells}
\title{\code{\link{skipWells}} rm wells from both data, plate maps and groupings}
\usage{
skipWells(data, skip)
}
\arguments{
\item{data}{data structures from \code{platexpress}; either data (\code{\link{readPlateData}}), a plate layout map (\code{\link{readPlateMap}}) or a well grouping (\code{\link{getGroups}})}

\item{skip}{a list of strings identifiying the wells to be skipped,
e.g. "B3" to skip the well in row B/column 3}
}
\description{
\code{\link{skipWells}} rm wells from both data, plate maps and groupings
}
\details{
removes specific wells from both data and groupins. If the first argument is \code{platexpress} data, the specified wells will be set to NA. If the first argument is a \code{platexpress} well grouping, the specified wells will be removed from the groups.
}
\examples{
data(ap12)
raw <- skipWells(ap12data, skip="A9")
}

