% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{skipWells}
\alias{skipWells}
\title{\code{\link{skipWells}} rm wells from both data, plate maps and groupings}
\usage{
skipWells(data, skip)
}
\arguments{
\item{data}{data structures from \code{\link{platexpress}}; either data
(\code{\link{readPlateData}}), a plate layout map
(\code{\link{readPlateMap}}) or a well grouping (\code{\link{getGroups}})}

\item{skip}{a list of strings identifiying the wells to be skipped,
e.g. "B3" to skip the well in row B/column 3}
}
\description{
\code{\link{skipWells}} rm wells from both data, plate maps and groupings
}
\details{
Removes specific wells from \code{\link{platexpress}} data,
groupings and plate layout maps. If the first argument is
\code{\link{platexpress}} data, the specified wells will be set to NA.
If the first argument is a \code{platexpress} well grouping or plate
layout map, the specified wells will be removed.
}
\examples{
data(ap12)
raw <- skipWells(ap12data, skip="A9") # rm data from well "A9"
plate <- skipWells(ap12plate, skip="A9") # rm well "A9" from the plate layout
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
