% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{viewGroups}
\alias{viewGroups}
\title{View Group Averages}
\usage{
viewGroups(data, yids, stats = "CI", groups, groups2, xid,
  xscale = FALSE, xlim, xlab, dtype = "data", ylab, pcols, group2.col,
  yscale = TRUE, ylims, ylim, log = "", lty.mean = 1, lwd.mean = 2,
  alpha = 55, lty.orig = 1, lwd.orig = 0.1, show.ci95 = FALSE,
  show.mean = TRUE, emphasize.mean = FALSE, g2.legpos = "topleft",
  g2.legend = TRUE, embed = FALSE, no.par = FALSE, mai = c(0.5, 0,
  0, 0), xmgp = c(1.5, 0.5, 0), ymgp = c(0, -1, -0.05), ytcl = 0.25,
  nrow = 1, xaxis = TRUE, yaxis = c(1, 2), g1.legpos = "topright",
  g1.legend = TRUE, verb = FALSE)
}
\arguments{
\item{data}{the platereader data object as provided by
\code{\link{readPlateData}}}

\item{yids}{IDs of the data to be plotted; if missing, all data will
be plotted}

\item{stats}{type of replicate statistics to plot, where replicates
are defined in option \code{groups2} or defined in the plate data object.
Options are either only a line for each replicate group ("mean" or
"median") or a line with transparent ranges. "CI": mean and 95\%
confidence interval (default), "SD": mean and standard deviation,
"SE": mean and standard error, "boxplot": median and 25\% and 75\%
quantiles, "range": median and full data range (min/max)}

\item{groups}{a list of well groups, wells from each
group will be plotted together in one plot, and statistics plotted
for sub-groups (replicates) in \code{groups2}. Groupings can also
be present as a list item in the \code{data} object.}

\item{groups2}{replicate well groups for which statistics are calculated
and plotted  (see argument \code{stats}), each group must be a subset
of exactly on well group in argument \code{groups}.}

\item{xid}{ID of a data-set in the input data that can be used as x-axis
instead of the default Time vector}

\item{xscale}{use a global range for the x-axes; only relevant if xid
specifies a subset of the data as x-axis}

\item{xlim}{plot range of the x-axis}

\item{xlab}{custom x-axis label}

\item{dtype}{type of the data to be plotted, default is the main 'data', but
e.g. 'orig' allows to plot the original data w/o processing (e.g.
interpolation, blank correction, etc.)}

\item{ylab}{custom y-axis label}

\item{pcols}{a named list of RGB colors to be used the plotted data types;
the color vector must have names according to the data IDs}

\item{group2.col}{a named list of RGB colors for group2 (which are shown
in one plot)}

\item{yscale}{if TRUE (default) global y-axis limits will be calculated from
all plotted wells; if FALSE each well be locally scaled}

\item{ylims}{a named list of y-axis ranges pairs for each data ID}

\item{ylim}{one y-axis limit range that will be used for all plotted data}

\item{log}{plot logarithmic axis, use equivalent to normal plot 'log', i.e.,
log="y" for a log y-axis, log="x" for x-axis and log="yx" for both axes}

\item{lty.mean}{line style for the mean lines, see \code{(par("lty")}
this can help to emphasize the means in noisy data (broad overlapping
confidence intervals}

\item{lwd.mean}{line width for the mean lines, see \code{(par("lwd")}}

\item{alpha}{hexadecimal alpha value for range plot opaqueness}

\item{lty.orig}{line type of the original single data, set to 0 to supress
plotting of all original data}

\item{lwd.orig}{line-width of the original single data, set to 0 to
supress plotting of all original data}

\item{show.ci95}{deprecated, but kept for backward compatibility, see
option \code{stats="CI"}: show the 95\% condifence intervals of groups;
set \code{stats="mean"} to reproduce previous \code{show.ci95=FALSE}
behaviour}

\item{show.mean}{show the mean of groups as lines}

\item{emphasize.mean}{show the means in black instead of group colors,}

\item{g2.legpos}{position of the \code{groups2} legend, see \code{g2.legend}}

\item{g2.legend}{plot a legend for groups in argument \code{groups2}}

\item{embed}{setting TRUE allows to embed plots of single groups within
in layouted plots, see ?layout and par("mfcol"); also see argument
\code{no.par}}

\item{no.par}{setting TRUE supresses all internal plot defaults (e.g.,
mai, mgp), useful to style your own plots, also see argument \code{embed}}

\item{mai}{set the outer margins around plot areas, see ?par}

\item{xmgp}{x-axis \code{mgp} settings: margin line in \code{mex} units
of axis title, tick labels and axis line; used only if \code{no.par==FALSE}}

\item{ymgp}{y-axis \code{mgp} settings: margin line in \code{mex} units
of axis title, tick labels and axis line}

\item{ytcl}{y-axis \code{tcl} settings: tick length and direction}

\item{nrow}{number of plot rows, number of columns will be selected
automatically; NOTE: to change the ordering of the plots
you can change the ordering of the input \code{groups}/}

\item{xaxis}{plot x-axis if TRUE}

\item{yaxis}{the data types for which axes are to be plotted, corresponds
to the order in argument \code{yids} and as plotted in the legend
(see argument \code{g1.legend}}

\item{g1.legpos}{position of the \code{groups} legend, see \code{g1.legend}}

\item{g1.legend}{show the main legend, giving plot colors}

\item{verb}{print progress messages}
}
\description{
plot grouped wells as summary plots, incl. means and confidence intervals.
This function gives a first overview of the reproducability between
replicates.
}
\examples{
data(ap12)
groups <- getGroups(plate=ap12plate, by=c("strain"))
vg <- viewGroups(ap12data,groups=groups,lwd.orig=0.1,nrow=1)
vg <- viewGroups(ap12data,groups2=groups,lwd.orig=0.1,nrow=1)
}
\seealso{
\code{\link{viewPlate}}, \code{\link{getGroups}},
\code{\link{readPlateMap}}
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
