% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{viewGroups}
\alias{viewGroups}
\title{plot grouped wells as summary plots, incl. confidence intervals and means}
\usage{
viewGroups(data, groups, mids = c(time = "Time", temp = "Temperature"), xid,
  xscale = FALSE, xlim, dids, pcols, yscale = TRUE, ylims, ylim, log = "",
  legpos = "topleft", lwd.orig = 0.5, mai = c(0.5, 0, 0, 0),
  mgp = c(1.5, 0.5, 0), nrow = 1, xaxis = TRUE, yaxis = TRUE)
}
\description{
plot grouped wells as summary plots, incl. confidence intervals and means
}
\seealso{
\code{\link{getGroups}}, \code{\link{readPlateMap}}
}

