% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{viewPlate}
\alias{viewPlate}
\title{View Plate}
\usage{
viewPlate(data, wells, wcols, rows = toupper(letters[1:8]), cols = 1:12,
  xid, xscale = FALSE, xlim, yids, dtype = "data", pcols, yscale = TRUE,
  ylims, ylim, log = "", legpos = "topleft", add.legend = TRUE)
}
\arguments{
\item{data}{the list of measurement data as provided by
\code{\link{readPlateData}}}

\item{wells}{a list of wells to plot, overrules \code{rows} and \code{cols}}

\item{wcols}{named color vector for wells, used for the well ID}

\item{rows}{a list of strings/characters used as row ID in the composite
row:col well description in the plate layout (map) and plate data}

\item{cols}{as rows but plate column IDs}

\item{xid}{ID of a data-set in the input data that can be used as x-axis
instead of the default Time vector}

\item{xscale}{use a global range for the x-axes; only relevant if xid specifies a subset of the data as x-axis}

\item{xlim}{plot range of the x-axis}

\item{yids}{IDs of the data to be plotted; if missing, all data will
be plotted}

\item{dtype}{type of the data to be plotted, default is the main 'data', but
e.g. 'orig' allows to plot the original data w/o processing (e.g.
interpolation, blank correction, etc.)}

\item{pcols}{a named list of RGB colors to be used the plotted data types; the color vector must have names according to the data IDs}

\item{yscale}{if TRUE (default) global y-axis limits will be calculated from
all plotted wells; if FALSE each well be locally scaled}

\item{ylims}{a named list of y-axis ranges pairs for each data ID}

\item{ylim}{one y-axis limit range that will be used for all plotted data}

\item{log}{plot logarithmic axis, use equivalent to normal plot 'log', i.e.,
log="y" for a log y-axis, log="x" for x-axis and log="yx" for both axes}

\item{legpos}{position of the well IDs on the plots}

\item{add.legend}{add a legend for the plotted data types (see
argument \code{yids}) in the last plotted well}
}
\description{
plots all data in plate format for a quick inspection of
the current data set. This function should be called immediately
after loading a new platereader dataset to inspect the data
and decide on further processing steps.
}
\examples{
data(ap12)
# view all data on the plate
viewPlate(ap12data) 
# inspect natural logarithm of OD_600 values, ie. log(X(t))
viewPlate(ap12data, yids="600", log="y") 
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
