% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereader.R
\name{viewPlate}
\alias{viewPlate}
\title{\code{\link{viewPlate}} plots all data in plate format}
\usage{
viewPlate(data, rows = toupper(letters[1:8]), cols = 1:12, mids = c(time =
  "Time", temp = "Temperature"), xid, xscale = FALSE, xlim, dids, pcols,
  yscale = TRUE, ylims, ylim, log = "", legpos = "topleft")
}
\description{
\code{\link{viewPlate}} plots all data in plate format
}
\examples{
data(ap12data)
vp <- viewPlate(ap12data)
}

